Moved the notifications out of state helper
